+++ packages/@markuplint/parser-utils/src/ignore-block.ts
@@ -1,6 +1,6 @@
 import type { Parser } from './parser.js';
 import type { Code, IgnoreBlock, IgnoreTag } from './types.js';
-import type { MLASTNodeTreeItem, MLASTPreprocessorSpecificBlock } from '@markuplint/ml-ast';
+import type { MLASTNodeTreeItem, MLASTPreprocessorSpecificBlock, MLASTText } from '@markuplint/ml-ast';
 
 import { MASK_CHAR } from './const.js';
 import { getPosition } from './get-location.js';
@@ -80,33 +80,106 @@ export function restoreNode(
 	}
 
 	for (const tag of stack) {
-		const node = newNodeList.find(node => node.startOffset === tag.index);
+		const raw = `${tag.startTag}${tag.taggedCode}${tag.endTag ?? ''}`;
+		const tagIndexEnd = tag.index + raw.length;
+
+		const node = newNodeList.find(node => node.startOffset <= tag.index && node.endOffset >= tagIndexEnd);
 
 		if (!node) {
 			continue;
 		}
 
-		const raw = `${tag.startTag}${tag.taggedCode}${tag.endTag ?? ''}`;
-		const token = parser.createToken(raw, node.startOffset, node.startLine, node.startCol);
+		const replacementChildNodes: (MLASTText | MLASTPreprocessorSpecificBlock)[] = [];
 
-		const psNode: MLASTPreprocessorSpecificBlock = {
-			...token,
-			type: 'psblock',
-			depth: node.depth,
-			nodeName: `#ps:${tag.type}`,
-			parentNode: node.parentNode,
-			childNodes: [],
-			isBogus: false,
-		};
+		if (node.startOffset === tag.index && node.endOffset === tagIndexEnd) {
+			const token = parser.createToken(raw, node.startOffset, node.startLine, node.startCol);
 
-		if (node.type !== 'doctype' && node.parentNode?.childNodes) {
-			parser.replaceChild(node.parentNode, node, psNode);
+			const psNode: MLASTPreprocessorSpecificBlock = {
+				...token,
+				type: 'psblock',
+				depth: node.depth,
+				nodeName: `#ps:${tag.type}`,
+				parentNode: node.parentNode,
+				childNodes: [],
+				isBogus: false,
+			};
+
+			replacementChildNodes.push(psNode);
+		} else if (node.type === 'text') {
+			const offset = tag.index - node.startOffset;
+			const above = node.raw.slice(0, offset);
+			const below = node.raw.slice(offset + raw.length);
+
+			if (above) {
+				const { line, column } = getPosition(node.raw, 0);
+				const token = parser.createToken(
+					above,
+					node.startOffset,
+					node.startLine + line - 1,
+					node.startCol + column - 1,
+				);
+
+				const aboveNode: MLASTText = {
+					...token,
+					nodeName: '#text',
+					type: 'text',
+					parentNode: node.parentNode,
+					depth: node.depth,
+				};
+
+				replacementChildNodes.push(aboveNode);
+			}
+
+			const { line, column } = getPosition(raw, offset);
+			const token = parser.createToken(
+				raw,
+				node.startOffset + offset,
+				node.startLine + line - 1,
+				node.startCol + column - 1,
+			);
+
+			const psNode: MLASTPreprocessorSpecificBlock = {
+				...token,
+				type: 'psblock',
+				depth: node.depth,
+				nodeName: `#ps:${tag.type}`,
+				parentNode: node.parentNode,
+				childNodes: [],
+				isBogus: false,
+			};
+			replacementChildNodes.push(psNode);
+
+			if (below) {
+				const { line, column } = getPosition(node.raw, offset + raw.length);
+				const token = parser.createToken(
+					below,
+					node.startOffset + offset + raw.length,
+					node.startLine + line - 1,
+					node.startCol + column - 1,
+				);
+
+				const aboveNode: MLASTText = {
+					...token,
+					nodeName: '#text',
+					type: 'text',
+					parentNode: node.parentNode,
+					depth: node.depth,
+				};
+
+				replacementChildNodes.push(aboveNode);
+			}
 		}
 
-		const index = newNodeList.indexOf(node);
-		newNodeList.splice(index, 1, psNode);
+		if (replacementChildNodes.length > 0) {
+			if (node.type !== 'doctype' && node.parentNode?.childNodes) {
+				parser.replaceChild(node.parentNode, node, ...replacementChildNodes);
+			}
 
-		tag.resolved = true;
+			const index = newNodeList.indexOf(node);
+			newNodeList.splice(index, 1, ...replacementChildNodes);
+
+			tag.resolved = true;
+		}
 	}
 
 	for (const node of newNodeList) {