+++ vscode/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 4.2.0
+
+- Depends: `markuplint@4.2.0`
+
+## 4.1.1
+
+- Depends: `markuplint@4.1.1`
+
 ## 4.1.0
 
 - Depends: `markuplint@4.1.0`